Add files via upload
I have developed this Fish Species Classification model using Machine Learning and Python. In this project, our objective is to predict the fish species considering the weight, height, length and width of the fishes. We will perform all the steps from Data gathering to Model deployment. During Model evaluation, we compare machine learning algorithms on the basis of accuracy score metric. Further, we will input the random data  including weight, height, length and width of the fish and ur model will predict the species of the fish.